backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
paho-mqtt==2.2.1
gunicorn==21.2.0